fix(desktop): VtEncoder bitrate_bps 参数为 u32 非 u64(按 vendored 签名核对)
本地不可编译 macOS 目标,预检 vendored vt_encoder.rs 签名发现
new_with_codec(bitrate_bps: u32)——u64 常量会在 macOS CI 类型错误,
先行修正。

+++ crates/aerodesk-desktop/src/macos_publisher.rs
@@ -25,7 +25,7 @@ use str0m::media::{Frequency, MediaTime};
 use str0m::net::Protocol;
 
 const FPS: u32 = 30;
-const DEFAULT_BITRATE_BPS: u64 = 8_000_000;
+const DEFAULT_BITRATE_BPS: u32 = 8_000_000;
 
 /// 当前发布线程的 stop 句柄（同一时刻只允许一个被控端发布线程）。
 static STOP: Mutex<Option<Arc<AtomicBool>>> = Mutex::new(None);
